Add recurring option to immediate task execution
Updated TaskManagerActionController and TaskQueue to support executing tasks immediately with a recurring flag. This allows tasks triggered for immediate execution to be optionally scheduled as recurring, improving flexibility in task management.
also fixes a bug that task manager keeps task as "Executing" when u connect to the account on other device (aka Recconect status)